update Makefile and README
This commit updates the Makefile to build Geant4 and ROOT automatically
by running the command:

    $ make install-deps

I also updated the README to explain how to compile everything locally,
and added a markdown version of the README for displaying on github.

I also fixed a small issue such that when the Results directory doesn't
exist, it gets created to avoid a segfault.