feat: First timing function implemented
* G4Args: Function for the storage of the 10 first photon times and
succesive sorting to incorporate the sortest ones among all detected
tracks.
Also a fuction to calculate the average btw both sides was created
TO BE DONE: The sorting algorithm can be improved not iterating through the entire array each time. Right now we can not modify the number of photons used for this calculation.

*util.cc: quicksort function

*event.cc: write out the global average time
*run.cc: included a row to store in the root file the global average
time