add: function to kill photons after n ps
* Introduced through -KillLT n where n is the number of ps.

TODO: use the G4OpticalPhoton description to kill only photons in this
manner. Right now I use a value of Edep>0 to be sure of that. The
definition was giving me issues. An estimate of maximum ps to reach both
SiPMs in an extreme case is 300 ps (the number should be higher than this one).