feat: FULL SYMMETRY MESH
The -GeomConfig 1 option generates the LYSO bar using a mesh of Znode*2
divisions along Z and 5 nodes per half cross section. We apply symmetry
to the ZX and ZY planes.

The bar is currently set lower than the nominal to test for light loss
if one of the triangles is not properly defined.

Next steps:
	Add commands and required values in G4Args for command line
execturion with Znode divisions and nodes. Be able to read non standard
radiuses per node
	Add option to use more than 5 nodes per cross section
	Further testing under non-standard radiuses