add: optional args modif & default rads
Changed input optional arguments as the numer of Z sections and the
number of nodes in the theta angle (TODO: modify this to a integer to
increase this number proportionally to Onode=5) and the pointer to the
array.
The vector is initialized after reading the input arguments in G4Args.
If the input optional argument is not null, it defines the vector using
the input data.