add: LYSO Volume Calculation Function
The volume of the LYSO mesh can be calculated through each triangular
face vertices. This provides a way to calculate a volume constraint for
the genetic optimizer outside the G4 simulation. The length of the LYSO
is currently fixed (TODO read -LYSO_L parameter in case this is changed,
optional parameter??)