add: Light Detection avg std + G4sim Getters
We compute now the light detection average and standard deviations per
run, even if it a single one. The results are given at the end of the
run, even if it is a single one.

These values can be obtained from the G4sim class with the right
Getters.

These values can be used for optimization purposes.